+++ src/wallet.cpp
@@ -36,7 +36,7 @@ static int64_t GetStakeCombineThreshold() { return 1 * COIN; }
 int64_t gcd(int64_t n,int64_t m) { return m == 0 ? n : gcd(m, n % m); }
 static inline uint64_t CoinWeightCost(const COutput &out)
 {
-    int64_t nTimeWeight = min((int64_t)GetTime() - (int64_t)out.tx->nTime, nStakeMaxAge);
+    int64_t nTimeWeight = min((int64_t)GetTime() - (int64_t)out.tx->nTime, (int64_t)nStakeMaxAge);
     CBigNum bnCoinDayWeight = CBigNum(out.tx->vout[out.i].nValue) * nTimeWeight / (24 * 60 * 60);
     return bnCoinDayWeight.getuint64();
 }
